Move `jive::Window` to `top-level/` subdirectory
