+++ jive_layouts/jive_layouts.cpp
@@ -20,12 +20,12 @@
 #include "layout/gui-items/flex/jive_FlexItem.cpp"
 #include "layout/gui-items/grid/jive_GridContainer.cpp"
 #include "layout/gui-items/grid/jive_GridItem.cpp"
+#include "layout/gui-items/top-level/jive_Window.cpp"
 #include "layout/gui-items/widgets/jive_Button.cpp"
 #include "layout/gui-items/widgets/jive_ComboBox.cpp"
 #include "layout/gui-items/widgets/jive_Label.cpp"
 #include "layout/gui-items/widgets/jive_ProgressBar.cpp"
 #include "layout/gui-items/widgets/jive_Slider.cpp"
-#include "layout/gui-items/widgets/jive_Window.cpp"
 
 #include "layout/gui-items/widgets/jive_Hyperlink.cpp"
 #include "layout/gui-items/widgets/jive_Knob.cpp"
