+++ jive_layouts/jive_layouts.h
@@ -45,12 +45,12 @@ namespace jive
 #include "layout/gui-items/flex/jive_FlexItem.h"
 #include "layout/gui-items/grid/jive_GridContainer.h"
 #include "layout/gui-items/grid/jive_GridItem.h"
+#include "layout/gui-items/top-level/jive_Window.h"
 #include "layout/gui-items/widgets/jive_Button.h"
 #include "layout/gui-items/widgets/jive_ComboBox.h"
 #include "layout/gui-items/widgets/jive_Label.h"
 #include "layout/gui-items/widgets/jive_ProgressBar.h"
 #include "layout/gui-items/widgets/jive_Slider.h"
-#include "layout/gui-items/widgets/jive_Window.h"
 
 #include "layout/gui-items/widgets/jive_Hyperlink.h"
 #include "layout/gui-items/widgets/jive_Knob.h"
